requirements.txt
mysql-connector-python==8.0.29
python-dotenv==0.20.0
pytest==7.1.2
